@@ -6,7 +6,7 @@ buildscript {
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
+        classpath 'com.android.tools.build:gradle:3.2.0'
         classpath 'com.github.dcendents:android-maven-gradle-plugin:2.0'
     }
 }
@@ -20,11 +20,13 @@ allprojects {
 }
 
 ext {
-    compileSdkVersion = 27
+    compileSdkVersion = 28
     buildToolsVersion = '28.0.0'
     targetSdkVersion = compileSdkVersion
 
-    supportLibVersion = '27.1.1'
+    appcompatLibVersion = '1.1.0'
+    cardViewLibVersion = '1.0.0'
+    materialLibVersion = '1.0.0'
     iconHandlerVersion = 'v2.0.0'
     butterKnifeVersion = '8.8.1'
 }
