@@ -1,6 +1,6 @@
 package com.vansuita.materialabout.util;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.view.View;
 
 /**
